+++ pkg/network/cache/clients.go
@@ -6,6 +6,7 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
+	"iter"
 	"maps"
 	"slices"
 	"sync"
@@ -79,9 +80,8 @@ func snCacheKey(pub []byte) string { return string(pub) }
 // Get initializes connections to network addresses of described SN and returns
 // interface to access them. All opened connections are cached and kept alive
 // until [Clients.CloseAll].
-func (x *Clients) Get(ctx context.Context, info clientcore.NodeInfo) (clientcore.MultiAddressClient, error) {
-	pub := info.PublicKey()
-	cacheKey := snCacheKey(pub)
+func (x *Clients) Get(ctx context.Context, info netmap.NodeInfo) (clientcore.MultiAddressClient, error) {
+	cacheKey := snCacheKey(info.PublicKey())
 
 	x.mtx.RLock()
 	c, ok := x.conns[cacheKey]
@@ -96,7 +96,7 @@ func (x *Clients) Get(ctx context.Context, info clientcore.NodeInfo) (clientcore
 		return c, nil
 	}
 
-	c, err := x.initConnections(ctx, pub, info.AddressGroup())
+	c, err := x.initConnections(ctx, info.PublicKey(), info.NetworkEndpoints())
 	if err != nil {
 		return nil, fmt.Errorf("init connections: %w", err)
 	}
@@ -138,21 +138,13 @@ func (x *Clients) syncWithNetmapSN(ctx context.Context, sn netmap.NodeInfo) erro
 		return nil
 	}
 
-	as := make(network.AddressGroup, 0, sn.NumberOfNetworkEndpoints())
-	var a network.Address
-	for ma := range sn.NetworkEndpoints() {
-		if err := a.FromString(ma); err != nil {
-			// TODO: if at least one address is OK, SN can be operational
-			return fmt.Errorf("parse network address %q: %w", ma, err)
-		}
-		as = append(as, a)
-	}
+	var as = slices.Collect(sn.NetworkEndpoints())
 
 	conns.mtx.Lock()
 	defer conns.mtx.Unlock()
 
 	maps.DeleteFunc(conns.m, func(ma string, c *client.Client) bool {
-		if slices.ContainsFunc(as, func(a network.Address) bool { return a.String() == ma }) {
+		if slices.Contains(as, ma) {
 			return false
 		}
 		if err := c.Close(); err != nil {
@@ -165,13 +157,12 @@ func (x *Clients) syncWithNetmapSN(ctx context.Context, sn netmap.NodeInfo) erro
 		return true
 	})
 
-	for i := range as {
-		ma := as[i].String()
+	for _, ma := range as {
 		if _, ok := conns.m[ma]; ok {
 			continue
 		}
 		x.log.Info("initializing connection to new SN address in the new network map...", zap.String("address", ma))
-		c, err := x.initConnection(ctx, pub, as[i].URIAddr())
+		c, err := x.initConnection(ctx, pub, ma)
 		if err != nil {
 			x.log.Info("failed to init connection to new SN address in the new network map",
 				zap.String("address", ma), zap.Error(err))
@@ -184,22 +175,21 @@ func (x *Clients) syncWithNetmapSN(ctx context.Context, sn netmap.NodeInfo) erro
 	return nil
 }
 
-func (x *Clients) initConnections(ctx context.Context, pub []byte, as network.AddressGroup) (*connections, error) {
-	m := make(map[string]*client.Client, len(as))
+func (x *Clients) initConnections(ctx context.Context, pub []byte, addrs iter.Seq[string]) (*connections, error) {
+	m := make(map[string]*client.Client)
 	l := x.log.With(zap.String("public key", hex.EncodeToString(pub)))
-	for i := range as {
-		cacheKey := as[i].String()
-		l.Info("initializing connection to the SN...", zap.String("address", cacheKey))
-		c, err := x.initConnection(ctx, pub, as[i].URIAddr())
+	for s := range addrs {
+		l.Info("initializing connection to the SN...", zap.String("address", s))
+		c, err := x.initConnection(ctx, pub, s)
 		if err != nil {
 			// TODO: if at least one address is OK, SN can be operational
 			for cl := range maps.Values(m) {
 				_ = cl.Close()
 			}
-			return nil, fmt.Errorf("init conn to %q: %w", as[i], err)
+			return nil, fmt.Errorf("init conn to %q: %w", s, err)
 		}
-		l.Info("connection to the SN successfully initialized", zap.String("address", cacheKey))
-		m[cacheKey] = c
+		l.Info("connection to the SN successfully initialized", zap.String("address", s))
+		m[s] = c
 	}
 	var hexKey = hex.EncodeToString(pub)
 	return &connections{
@@ -210,7 +200,13 @@ func (x *Clients) initConnections(ctx context.Context, pub []byte, as network.Ad
 }
 
 func (x *Clients) initConnection(ctx context.Context, pub []byte, uri string) (*client.Client, error) {
-	target, withTLS, err := uriutil.Parse(uri)
+	// FIXME: pending removal in #3982.
+	var a network.Address
+	if err := a.FromString(uri); err != nil {
+		return nil, fmt.Errorf("parse network address %q: %w", uri, err)
+	}
+
+	target, withTLS, err := uriutil.Parse(a.URIAddr())
 	if err != nil {
 		return nil, fmt.Errorf("parse URI: %w", err)
 	}